320 - First refactor
