+++ app/app/play/page.tsx
@@ -1,7 +1,7 @@
 import { currentUser } from '@clerk/nextjs/server';
 import { redirect } from 'next/navigation';
-import { Game } from '~/components/game/game';
 import { GameFinish } from '~/components/game/game-finish';
+import { UserGame } from '~/components/usergame/usergame';
 import { getRandom, getUnplayedScenarios } from '~/lib/db/queries';
 
 export default async function Page() {
@@ -25,5 +25,5 @@ export default async function Page() {
         redirect('/app/scores');
     }
 
-    return <Game scenario={scenario} unplayedScenarios={unplayedScenarios.length} />;
+    return <UserGame scenario={scenario} unplayedScenarios={unplayedScenarios.length} />;
 }
