+++ app/backstage/play/[scenarioId]/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from 'next/navigation';
 import { getScenario } from '~/lib/db/queries';
-import { Game } from '~/components/game/game';
+import { UserGame } from '~/components/usergame/usergame';
 
 export default async function Page({ params }: { params: Promise<{ scenarioId: number }> }) {
     const id = (await params).scenarioId;
@@ -10,5 +10,5 @@ export default async function Page({ params }: { params: Promise<{ scenarioId: n
 
     if (!scenario?.data) notFound();
 
-    return <Game backstageAccess scenario={scenario} />;
+    return <UserGame backstageAccess scenario={scenario} />;
 }
